pyproject.toml
[project]
name = "mblabelsite"
version = "1.0.0"
description = "Static site generator for tici taci records, powered by MusicBrainz"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "musicbrainzngs>=0.7",
    "click>=8.0",
    "Pillow>=10.0",
    "pydantic>=2.0",
]

[project.scripts]
mblabelsite = "mblabelsite.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mblabelsite"]

[tool.commitizen]
version = "1.0.0"
version_files = ["__version__.py", "pyproject.toml:version"]
update_changelog_on_bump = true

[tool.mblabelsite]
label_id = "62db3e96-423a-4e9d-bf66-7a017f1dfc73"
site_url = "https://ticitaci.com"
excluded_release_ids = ["14f0377a-dadf-4fb9-a141-4a6e8c9ed882"]
ignored_artist_ids = [
    "8a26ca9b-d542-449b-a5e7-224da9eb8a77",
    "89ad4ac3-39f7-470e-963a-56509c546377",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.8"]
